+++ routers/conversation_router.py
@@ -1,10 +1,4 @@
 from mira import (
-    status,
-    context_processor,
-    audio_scorer,
-    wake_word_detector,
-    command_processor,
-    inference_processor,
     logger,
 )
 from db import get_db_session
