+++ routers/interaction_router.py
@@ -1,6 +1,5 @@
-from mira import status, context_processor, audio_scorer, wake_word_detector, command_processor, inference_processor, logger
+from mira import status, context_processor, audio_scorer, wake_word_detector, command_processor, inference_processor, SentenceProcessor, logger
 from db import get_db_session
-import processors.sentence_processor as SentenceProcessor
 from models import Interaction, Person
 from fastapi import APIRouter, UploadFile, File, Form, HTTPException
 import uuid
